next_word {NxT}	R Documentation

Ngram Model to predict next word We built and train three ngram to check what will be the next word, we check first with the last 3 words, if nothing is found, the last two and so the last. If there is no match, the word the most used is returned.

Description

Ngram Model to predict next word We built and train three ngram to check what will be the next word, we check first with the last 3 words, if nothing is found, the last two and so the last. If there is no match, the word the most used is returned.

Usage


next_word(str1)

Arguments

str1
a sentence or word, just the maximum last three words will be in the process.

Details

This function is the last test for the data science coursera course of Hopkins university

Value

the five word with the higher frequency following the precedents words.

Author(s)

Aurelien Roussille

See Also

filter, tolower,str_split

[Package NxT version 1.0 Index]